Update root node id during node creation in Android accessibility
This CL removes a check that prevents the mCurrentRootId from updating
during the creation of AccessibilityNodeInfo objects.

With the current behavior, in some cases Android-side code is not
receiving a signal that the page was navigated, and so the root ID is
not being updated. When the root ID is stale, then we provide a stale
AccessibilityNodeInfo tree to downstream clients/AT. With this change
we always fetch a fresh root ID during node creation to eliminate the
possibility of a stale tree, at the cost of one extra JNI trip (which
is minimal/trivial). This will keep the mCurrentRootId up-to-date with
the native browser-side AXTree. This can be considered a stopgap
solution to a P1 bug, and caching can be revisited in the future, but
a cache of this value may not be needed in general.

AX-Relnotes: N/A
Bug: 396447488
Change-Id: I134d2313c68c04973aefdaee82bd181ba1a47293
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6302633
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#1425146}