[PEPC] Fix CHECK failure in payment_handler_permission_context.cc
This CL addresses a crash in payment_handler_permission_context.cc,
which was reported by lib fuzzer. The crash occurs when the browser
receives a PEPC request for an unsupported type (likely due to auto
generated mojo inputs). To prevent this, the CL adds checks to ensure
that the browser only handles PEPC requests of supported types.

Bug: 398356454
Change-Id: I8d9de086486066876b5d8d07674ec94dd791b51e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6298138
Commit-Queue: Yifan Luo <lyf@chromium.org>
Reviewed-by: Thomas Nguyen <tungnh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425158}