Introduce AutofillPredictionSource to AutofillField.
This CL also introduces an internal PredictionResults struct so that we
store the cache in one place.

As per the tests: I'd argue assigning correct prediction source is the
part of the same behaviour that assigns the type therefore I extended
the existing tests instead of adding separate one.

This is a new approach to
https://chromium-review.googlesource.com/c/chromium/src/+/6163744
(effectively replacing it) and a deviation from
go/field-prediction-overlap-metric-design.

Bug: 376432267
Change-Id: I6ee20bc9c22572afbade008261433369dbeae3de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6253099
Reviewed-by: Dominic Battré <battre@chromium.org>
Commit-Queue: Piotr Domaradzki <doland@google.com>
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1425184}