[Protected Audiences] Pass FrameTreeNodeIds to KVv2 cache path.
And key the cache on these IDs. These will be used to add events for
cache-managed network requests to devtools for the frame.

The ID is added unconditionally as part of the FetchKey, which
makes it so different frames can't share KVv2 cache entries, but I hope
to relax that constraint in the future (Could remove it from the key
on fetch complete, or only populate it when multiple frames request
related resources, or both, etc).

Bug: 379843989
Change-Id: I3e18b49341b928057cbfd855fc3dabb5dd0cb706
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6300675
Commit-Queue: mmenke <mmenke@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425192}