Cleanup Canvas text direction tests
Fix various style issues. And add tests for default, valid and invalid
values. The tests for valid and default values fail because we never
report `inherit`, but report the computed value instead.

A spec issue was filed for the round-tripping of `direction` and it was
resolved to return "inherit" rather than the computed value. See
https://github.com/whatwg/html/issues/10884 and the spec PR at
https://github.com/whatwg/html/pull/10873.

The code will not be changed to pass these tests until the results of
UseCounters have been reviewed to identify how often this change might
affect sites. A search of HTTP Archive did not find failure cases,
mostly because Firefox already returns "inherit" so sites check for
that.

Change-Id: I39645acd997786c22d4983f84a9d17c7b976797b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6277644
Reviewed-by: Andres Ricardo Perez <andresrperez@chromium.org>
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425228}