Factor out download protection functionality into delegate interface
This adds a DownloadProtectionDelegate interface which will have
separate implementations for desktop and Android. The desktop
implementation is copied from the existing DownloadProtectionService
code. An implementation for Android will be added in a later CL.

As a temporary measure, the desktop implementation of the delegate is
instantiated on Android, but only in tests.

Also updates the network traffic annotation for client_download_request.
There will soon be different completions for desktop and Android
versions of this network traffic annotation.

Bug: 397407934
Change-Id: Iad0a6c32a68547a1259aaa1704e2ed8b5b0f4151
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6301299
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Commit-Queue: Lily Chen <chlily@chromium.org>
Reviewed-by: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425229}