Remove the originating saved group after share and unshare
When unsharing a shared group, the originating saved group should
be removed. This is because once the shared group is no longer
accessible, the originating saved group will no longer be filtered
out. To solve this, it should be removed  immediately on successful unshare.

Bug: 399200395
Change-Id: Iaa072ca6f999ea742f0fbe993df8f1e2da29b980
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6302665
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Reviewed-by: Rushan Suleymanov <rushans@google.com>
Commit-Queue: Min Qin <qinmin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425274}