Avoid string copies of JSON map keys that contain no escape sequences
When a map key contains no escape sequences, it is possible for serde to
provide a str borrowed from the input, rather than allocating a new
string. Logically we could therefore call next_key::<Cow<str>> to take
advantage of this, since all we need to do is provide a rust::Str back
to C++.

Unfortunately, the blanket Deserialize impl for Cow<T> cannot avoid
copies of the borrowed str without specialization, so instead we provide
a wrapper type that manually implements this.

Change-Id: I2959d1f5f5251552ec0f5753a9f731e6052a68aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6305600
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425278}