Override visited link functions to avoid test crashes
The PhishingClassifierBrowserTest and
PhishingDomFeatureExtractorBrowserTest both rely on a mock
ChromeContentRendererClient for their test cases. However, it does not
initialize the visited_link_reader_ member, so any calls attempting to
determine a link's :visited-ness will crash the test with a null-pointer
dereference. In the past, this test has just overridden these functions
inside the test suite and had them return 0 or false. Since the
introduction of partitioned :visited links, more functions have been
introduced to ContentRendererClient which rely on a valid
visited_link_reader_. This CL overrides those new functions to avoid
crashes in the future, once the feature is enabled.

Change-Id: I5b33485b4ca003e1a5e9bfb2c2842bce11f2de58
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6305386
Commit-Queue: Kyra Seevers <kyraseevers@chromium.org>
Reviewed-by: Nathan Parker <nparker@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425310}