Translator: Gate new V1 behavior on TranslationAPIV1
Currently availability may report that translation pair is downloadable
when it is not. This is because availability is implemented with V1
behavior whereas create is implemented with OT behavior.

This CL updates it so that when TranslationAPIV1 is off, both
availability and create have OT behavior. When it is on, they have V1
behavior.

Bug: 385173766
Change-Id: I36fe9f22d8388790773e25b36744cf5249e70a6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6297840
Commit-Queue: Nathan Memmott <memmott@chromium.org>
Reviewed-by: Ming-Ying Chung <mych@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425324}