Speculative fix #2 for crash in media settings view
In the ongoing search for why QuickSettingsMediaView::HideItem crashes
occasionally, this CL makes item reordering robust against a different
number of ordered items vs total items.

Specifically, when ordering, only active controllable items are included
in the list of ordered ids. However, there can be other items in the
Quick Settings view that are frozen. In the previous code, all `items_`
were removed from the view tree and then the ordered (sub)set was
re-added, but `items_` still held raw ptrs to the
soon-to-be-deleted-when-the-fn-exits set difference.

This CL changes the behavior of the reordering code so that it no longer
frees any items. Instead, it simply moves the ordered items to the front
of the container view's child list. Any set difference ends up at the
end of the container, rather than dropped.

Will try to write a test for this as a follow-up, but repro is somewhat
tricky. The easiest approach is to have multiple fg tabs, each of which
creates / destroys media players (e.g., set `src= ""; src="..."; await
play()`). This will request audio focus, and lose focus for the other
tabs. When focus is lost, the set of active/controllable items can
differ from the total set of items. If one opens the shelf manually at
that point, it can enter with the size of `ids` less than `items_`.

Bug: 368135848
Change-Id: Ib0f7933cc43e110316b30fca3323ba490257a069
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6302327
Commit-Queue: Frank Liberato <liberato@chromium.org>
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425338}