Fork grpc instead of importing it as a submodule
Note to reviewers: don't review third_party/grpc/sources/, review
the rest and verify that update.sh is a no-op locally.

Rolling gRPC to a more recent revision in https://crrev.com/c/6276328
will require Chromium to apply some patches. This CL forks gRPC to
add the basis for that. For now we still use the same pinned gRPC
revision, so there should be no behavior change.

The fork strategy is the same used for abseil and mediapipe, i.e. the
files are checked into the src/ repository. To roll, update.sh
downloads the new revision and applies the patches in the patches/
directory (there are none in this CL yet). The script avoids checking
unnecessary files into the repo such as tests and non-C++ files.
generate_gn.sh does depend on non-C++ files, so it's updated to depend
on a fresh clone of the repo (at the same revision).

The removed submodule used to live in third_party/grpc/src. Sadly, this
CL cannot apply cleanly if the fork uses the same path, so it lives
in third_party/grpc/source instead. Includes and BUILD.gn are updated
accordingly.

NO_IFTTT=False positives inside the forked repo.

Bug: 40285204
Change-Id: I3bc2ac38cacdbb8b3527984a47f331dabe6f0aa7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6269394
Reviewed-by: Nico Weber <thakis@chromium.org>
Owners-Override: Nico Weber <thakis@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Vigen Issahhanjan <vigeni@google.com>
Cr-Commit-Position: refs/heads/main@{#1425352}