[css-env] Add env(preferred-text-scale)
This env variable exposes the text scale as chosen by the user in the OS
accessibility settings.

The name is not necessarily final. We will change it when the WG is
finished bikeshedding.

This is currently enabled with experimental web platform features or
with the CSSPreferredTextScale flag.

Bug: 397737223
Change-Id: Iefbcde80371c26e7f72a87740dcf07a81bce945c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6299839
Commit-Queue: David Grogan <dgrogan@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425413}