Video Effects: switch from indices to tags
It is better to use named tags when specifying calculator inputs
(as opposed to indices) for MediaPipe calculators. This CL makes
each calculator expose named tags that it expects for its inputs,
and switches the graph creation to use them.

Additionally, background blur calculator does not use StaticConfig,
so this input was removed from it.

Change-Id: Ie371d6d5ab39e7ea4198b797798fad2886c97d1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6307222
Reviewed-by: Bryant Chandler <bryantchandler@chromium.org>
Commit-Queue: Piotr Bialecki <bialpio@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425429}