[TabDragDrop] Pass window ID to retrieve tab by window
Build windowId into ChromeDropDataAndroid so that DragAndDropLauncherActivity can store the source windowId in the intent when #buildClipData. This allows us to use the source windowId in ChromeTabbedActivity to find the tab by its window. This way we can avoid looping through all windows which can be expensive if we can have more than 5 windows in LFF.

Bug: 398262317
Change-Id: Iff0ee952dbf81dbc4eebe2c04e8b1d66964d2f34
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6299073
Reviewed-by: Sky Malice <skym@chromium.org>
Commit-Queue: Zhe Li <zheliooo@google.com>
Cr-Commit-Position: refs/heads/main@{#1425450}