Create a new header for gardener rotation consoles
This creates a new DEV_HEADER to be used on the new consoles that
we want to incorporate into the existing header for inspection.
This also creates a "Tree Closers" console which is likely way too
large to be of use but I believe visualizing it will help make it more
obvious how this should be laid out (maybe split it into "gardener
rotation" tree closers). It also renames the rotation console names.

Bug: 40873502
Change-Id: Ica4a17aa34396e3cc54f5e17512cb1d9d0d8d23b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6302580
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Cr-Commit-Position: refs/heads/main@{#1425480}