Extend Autofill Private API to provide more info about entities
This CL adds 3 new methods to the API that will be used in
chrome://settings/autofillAi.

`getEntityInstanceByGuid` returns a complete entity instance by its
guid. This will be called when a user clicks on the "Edit entity
instance" button. Upon click, the entity instance will be retrieved by
its guid, so that the pre-existing values of its attributes are
populated in the form.

`getAllEntityTypes` provides a list of all possible entity types, along
with some strings for each entity. This will be called when the user
clicks on the "Add new entity instance" button. Upon click, a list of
all entity types will be shown, and the user has to choose one of them.

`getAllAttributeTypesForEntity` provides a list of all attributes an
entity can have. This will be called when the user adds or edits an
entity instance, in order to display all possible attributes that the
user can set.

Bug: 393318055
Change-Id: Iee3a268cd0cd5b5baaa303f5a8a28b8372e46f88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6244464
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Reviewed-by: Justin Lulejian <jlulejian@chromium.org>
Commit-Queue: Theo Cristea <theocristea@google.com>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425600}