[iOS] Hide passkey upgrades toggle on saving disabled
Modifies the logic of settings VC slightly to store the value of saving
passkeys pref instead of whether it's managed (the latter doesn't really
matter, the behaviour of the toggle should only be affected by whether
saving passkeys is disabled).

The toggle is now visible when both saving (passkeys, passwords) prefs
are enabled and the automatic passkey upgrades feature is enabled;
otherwise it's hidden.

Bug: 358343061
Change-Id: I5e5bca44a17f63a1d7d7c483951a24afbfed5551
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6298119
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Commit-Queue: Rafał Godlewski <rgod@google.com>
Cr-Commit-Position: refs/heads/main@{#1425643}