[Passwords] Refactor check for manual generation offer in FormParserTest
Current test setup is sub-optimal, because we can only check if
manual generation is offered on a field, but cannot check that it is
not offered.
Also treating manual generation as an "ElementRole" (username, current
password, new password) is semantically confusing, because a field
can be a field with a manual generation offering, and e.g. a new
password field at the same time.
Fixed those issues.

This is a purely mechanical change, no tests are added or removed.

Change-Id: I9bac07ba4ae5d0d1cb4c56b2b26cb56177a814f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6304629
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Commit-Queue: Maria Kazinova <kazinova@google.com>
Cr-Commit-Position: refs/heads/main@{#1425679}