Add 'See more' integration for Price Tracking card
This CL makes 'See more' visible for the Price Tracking card and
integrates it with the price tracking menu. There is more to be done -
for example, the menu will need to be customized from this entry point
(not 'current page' for example). But this CL lays the foundation.

Bug: 392972945
Change-Id: Ie5f6f80f5b28ed8eede2ea680fc30016e7308bc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6303021
Commit-Queue: David Maunder <davidjm@chromium.org>
Reviewed-by: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1425700}