win: open glic detached when browser not in foreground
Instead of open glic detached when the browser window is occluded,
check that the browser window is either active or the top most
visible, occluding window, using the native window occlusion code
to filter out various windows (cloaked, invisible, floating, etc).

This entailed moving the functionality that determines if a window
is visible and occluding to gfx/win/hwnd_util, so that it can be
shared between the occlusion code and the glic code.

Bug: 396733553
Change-Id: I43ad2ffc20cd1365e5f5bc44ad98f6b8bfcbd892
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6300097
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: David Bienvenu <davidbienvenu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425713}