Drop ref to shared image after Dissociate
Originally WebGPUMailboxTexture was holding Mailbox which doesn't keep
shared image alive on its own. As part of ClientSharedImage refactoring
it was switched to use ClientSharedImage in https://crrev.com/c/6017025

ClientSharedImage is owning reference and extends life-time of the
underlying memory.

There is evidence that in some cases WebGPUMailboxTexture is not
destroyed in a timely manner, which leads to increased memory pressure.
While we're resolving that we should drop reference to
ClientSharedImage after DissociateMailbox.

We never use `shared_image_` after it was called and it implicitly
kept alive on the service side before it because we have open access
on it.

Bug: 384096059
Change-Id: Ieffd4344597206df668f60637bad9ec50c620ded
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6308963
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425718}