Provide an audience value which includes the scheme
Today we are only generating JWTs where we will be the ones validating
the token and payload so the exact format of the audience param isn't
that important. We may decide to use the JWT for API auth in the
future though so we may as well use an audience format which will
work for either case.

Bug: 388885661
Change-Id: Ie813a2e2b17abb8cb4436e68acd7777a10c46624
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6306135
Reviewed-by: Yuwei Huang <yuweih@chromium.org>
Commit-Queue: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425723}