[CodeHealth] Clean up the ContentCapture feature
Since the ContentCapture has been enabled by default for Android and
disabled by default for other platforms for a while, this CL removes
the feature and updates IsContentCaptureEnabled() to return true for
Android and false otherwise.

Additionally, since ContentCaptureTriggeringForExperiment has been
disabled by default for a while, this CL removes the feature and all
the code that uses it.

Bug: 356236497, 356236267
Change-Id: I07589324090002727a03efa979139f0c5ac7a1bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6306395
Commit-Queue: Oksana Zhuravlova <oksamyt@chromium.org>
Reviewed-by: Peilin Wang <peilinwang@google.com>
Cr-Commit-Position: refs/heads/main@{#1425739}