[GAS] Prevent deleting duplicates during updates
If the feature `kAutofillDeduplicateAccountAddresses` is enabled then
exact duplicates that occurred due to an update are not deleted.

This is needed since the deduplication logic calls
UpdateProfile(Merge(a, b) = a), RemoveProfile(b), in this order.
As a result, it triggers a user-initiated removal of b. By the time
the silent removal from the deduplication code is triggered, it’s too
late.

Additionally this CL fixes a typo in the flag name.

Bug: 357074792
Change-Id: I5b35f9f96ab519fddd0b18da94dde08bb0bfa21a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6304602
Commit-Queue: Karol Sygiet <sygiet@google.com>
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1425767}