[a11y] Lower threshold for non-stopword image alt text to 3 characters
This commit lowers the threshold for image alt
text from 4 to 3 characters, but only for alt text
that is not composed primarily of stop words. This
means that images with alt text that is 3
characters or longer and contains non-stop words
will no longer be considered for automatic image
descriptions.

This change is being made in response to user
feedback that the previous threshold was too high
and was preventing some images with useful alt
text from being described. Because this change
only affects alt text containing non-stop words,
its impact is expected to be low.

Alternatives considered:

* Removing the threshold altogether. This was rejected because it could
  lead to generating image descriptions for images where they are not needed,
  potentially creating a confusing experience for users.

* Modifying the screen reader hint to provide different guidance when:
    * An image has no alt text at all (suggesting
      the use of the context menu to get a
      description).
    * An image has some alt text present
      (suggesting the use of the context menu for
      expanded image descriptions).

The following histograms will be used to track the impact of this change:

* Accessibility.ImageAnnotation.ServerRequests
* Accessibility.ImageAnnotation.CacheHits

If we see a significant decrease in the number of
server requests or cache hits, we may need to
revisit this change.

Bug: 389897606

Change-Id: Ife64a1bfd7cd1be06c174a02338ec7e9065a1b6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6191348
Reviewed-by: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Lucas Radaelli <lucasradaelli@google.com>
Cr-Commit-Position: refs/heads/main@{#1425794}